Rename LimitedWrite to Limited
And move it to the io module in futures-test.
