+++ futures-test/src/io/mod.rs
@@ -1,6 +1,7 @@
 //! Additional combinators for testing async IO.
 
 mod interleave_pending;
+mod limited;
 
 pub mod read;
 pub use read::AsyncReadTestExt;
